Split iodata into iodata and api modules
This way iodata can be imported from the format modules for proper type
checking.
