+++ iodata/__init__.py
@@ -19,4 +19,5 @@
 """Input and Output Module."""
 
 
-from .iodata import *
+from .iodata import IOData
+from .api import *
