+++ iodata/iodata.py
@@ -19,33 +19,12 @@
 """Module for handling input/output from different file formats."""
 
 
-import os
-from typing import List, Tuple, Type, Iterator
-from types import ModuleType
-from fnmatch import fnmatch
-from pkgutil import iter_modules
-from importlib import import_module
+from typing import List, Tuple, Type
 
 import numpy as np
 
-from .utils import LineIterator
 
-
-__all__ = ['IOData', 'load_one', 'load_many', 'dump_one', 'dump_many']
-
-
-def find_format_modules():
-    """Return all file-format modules found with importlib."""
-    result = {}
-    for module_info in iter_modules(import_module('iodata.formats').__path__):
-        if not module_info.ispkg:
-            format_module = import_module('iodata.formats.' + module_info.name)
-            if hasattr(format_module, 'patterns'):
-                result[module_info.name] = format_module
-    return result
-
-
-format_modules = find_format_modules()
+__all__ = ['IOData']
 
 
 class ArrayTypeCheckDescriptor:
@@ -417,138 +396,3 @@ class IOData:
             self._spinpol = spinpol
         else:
             raise TypeError("spinpol cannot be set when orbitals are present.")
-
-
-def _select_format_module(filename: str, attrname: str, fmt: str = None) -> ModuleType:
-    """Find a file format module with the requested attribute name.
-
-    Parameters
-    ----------
-    filename
-        The file to load or dump.
-    attrname
-        The required atrtibute of the file format module.
-    fmt
-        The name of the file format module to use. When not given, it is guessed
-        from the filename.
-
-    Returns
-    -------
-    format_module
-        The module implementing the required file format.
-
-    """
-    basename = os.path.basename(filename)
-    if fmt is None:
-        for name, format_module in format_modules.items():
-            if any(fnmatch(basename, pattern) for pattern in format_module.patterns):
-                if hasattr(format_module, attrname):
-                    return format_module
-    else:
-        return format_modules[fmt]
-    raise ValueError('Could not find file format with feature {} for file {}'.format(
-        attrname, filename))
-
-
-def load_one(filename: str, fmt: str = None) -> IOData:
-    """Load data from a file.
-
-    This function uses the extension or prefix of the filename to determine the
-    file format. When the file format is detected, a specialized load function
-    is called for the heavy lifting.
-
-    Parameters
-    ----------
-    filename
-        The file to load data from.
-    fmt
-        The name of the file format module to use. When not given, it is guessed
-        from the filename.
-
-    Returns
-    -------
-    out
-        The instance of IOData with data loaded from the input files.
-
-    """
-    format_module = _select_format_module(filename, 'load_one', fmt)
-    lit = LineIterator(filename)
-    try:
-        return IOData(**format_module.load_one(lit))
-    except StopIteration:
-        raise lit.error("File ended before all data was read.")
-
-
-def load_many(filename: str, fmt: str = None) -> Iterator[IOData]:
-    """Load multiple IOData instances from a file.
-
-    This function uses the extension or prefix of the filename to determine the
-    file format. When the file format is detected, a specialized load function
-    is called for the heavy lifting.
-
-    Parameters
-    ----------
-    filename
-        The file to load data from.
-    fmt
-        The name of the file format module to use. When not given, it is guessed
-        from the filename.
-
-    Yields
-    ------
-    out
-        An instance of IOData with data for one frame loaded for the file.
-
-    """
-    format_module = _select_format_module(filename, 'load_many', fmt)
-    lit = LineIterator(filename)
-    for data in format_module.load_many(lit):
-        try:
-            yield IOData(**data)
-        except StopIteration:
-            return
-
-
-def dump_one(iodata: IOData, filename: str, fmt: str = None):
-    """Write data to a file.
-
-    This routine uses the extension or prefix of the filename to determine
-    the file format. For each file format, a specialized function is
-    called that does the real work.
-
-    Parameters
-    ----------
-    iodata
-        The object containing the data to be written.
-    filename
-        The file to write the data to.
-    fmt
-        The name of the file format module to use. When not given, it is guessed
-        from the filename.
-
-    """
-    format_module = _select_format_module(filename, 'dump_one', fmt)
-    with open(filename, 'w') as f:
-        format_module.dump_one(f, iodata)
-
-
-def dump_many(iodatas: Iterator[IOData], filename: str, fmt: str = None):
-    """Write multiple IOData instances to a file.
-
-    This routine uses the extension or prefix of the filename to determine
-    the file format. For each file format, a specialized function is
-    called that does the real work.
-
-    Parameters
-    ----------
-    iodatas
-        An iterator over IOData instances.
-    filename : str
-        The file to write the data to.
-    fmt
-        The name of the file format module to use.
-
-    """
-    format_module = _select_format_module(filename, 'dump_many', fmt)
-    with open(filename, 'w') as f:
-        format_module.dump_many(f, iodatas)
