+++ iodata/test/test_gaussianlog.py
@@ -21,7 +21,7 @@
 
 from numpy.testing import assert_equal, assert_allclose
 
-from ..iodata import load_one
+from ..api import load_one
 
 try:
     from importlib_resources import path
