+++ iodata/test/test_orca.py
@@ -22,7 +22,7 @@
 import numpy as np
 from numpy.testing import assert_equal, assert_allclose
 
-from ..iodata import load_one
+from ..api import load_one
 from ..utils import angstrom
 
 try:
