+++ iodata/test/test_poscar.py
@@ -24,8 +24,8 @@ import os
 import numpy as np
 from numpy.testing import assert_equal, assert_allclose
 
+from ..api import load_one, dump_one
 from ..utils import angstrom, volume
-from ..iodata import load_one, dump_one
 
 try:
     from importlib_resources import path
